+++ lmu/policy/base/browser/configure.zcml
@@ -1,0 +1,100 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:browser="http://namespaces.zope.org/browser"
+    xmlns:i18n="http://namespaces.zope.org/i18n"
+    i18n_domain="lmu.policy.base">
+
+    <browser:page
+        name="search"
+        class=".Search"
+        permission="zope2.View"
+        for="*"
+        template="templates/search.pt"
+        layer="lmu.policy.base.interfaces.ILMUBaseThemeLayer"
+        />
+
+    <browser:page
+        for="*"
+        name="livesearch_reply2"
+        class="lmu.policy.base.browser.LivesearchReply"
+        permission="zope2.View"
+        layer="lmu.policy.base.interfaces.ILMUBaseThemeLayer"
+        />
+
+    <browser:page
+        for="*"
+        name="user_info.include"
+        class="lmu.policy.base.browser.UserInfo"
+        permission="zope2.View"
+        layer="lmu.policy.base.interfaces.ILMUBaseThemeLayer"
+        />
+
+    <browser:viewlet
+        name="lmu.path_bar"
+        manager="plone.app.layout.viewlets.interfaces.IAboveContent"
+        class="lmu.policy.base.browser.PathBarViewlet"
+        layer="lmu.policy.base.interfaces.ILMUBaseThemeLayer"
+        permission="zope2.View"
+        />
+
+    <browser:page
+        name="content_view"
+        for="lmu.policy.base.interfaces.ILMUContent"
+        permission="cmf.ModifyPortalContent"
+        class=".content.EntryContentView"
+        />
+
+    <browser:page
+        name="content_sort_files_view"
+        for="lmu.policy.base.interfaces.ILMUContent"
+        permission="cmf.ModifyPortalContent"
+        class=".content.EntrySortFilesView"
+        />
+
+    <browser:page
+        name="content_sort_images_view"
+        for="lmu.policy.base.interfaces.ILMUContent"
+        permission="cmf.ModifyPortalContent"
+        class=".content.EntrySortImagesView"
+        />
+
+    <browser:page
+        for="plone.app.contenttypes.interfaces.IFile"
+        name="contained_file_edit"
+        permission="cmf.ModifyPortalContent"
+        class=".content.ContainedFileEditForm"
+        />
+
+    <browser:page
+        for="plone.app.contenttypes.interfaces.IImage"
+        name="contained_image_edit"
+        permission="cmf.ModifyPortalContent"
+        class=".content.ContainedImageEditForm"
+        />
+
+    <browser:page
+        for="plone.app.discussion.interfaces.IComment"
+        name="add"
+        permission="cmf.ModifyPortalContent"
+        class=".content.LMUCommentAddForm"
+        />
+
+    <browser:page
+        name="ploneform-render-widget"
+        for="z3c.form.interfaces.IWidget"
+        layer="lmu.policy.base.interfaces.ILMUCommentFormLayer"
+        class=".content.LMURenderWidget"
+        permission="zope.Public"
+        />
+
+    <browser:viewlet
+        name="plone.comments"
+        for="Products.CMFCore.interfaces.IContentish"
+        layer="lmu.policy.base.interfaces.IProductLayer"
+        manager="plone.app.layout.viewlets.interfaces.IBelowContent"
+        view="plone.app.layout.globals.interfaces.IViewView"
+        class=".content.LMUCommentsViewlet"
+        permission="zope2.View"
+        />
+
+</configure>
