+++ lmu/policy/base/browser/templates/entry_edit.pt
