+++ lmu/policy/base/browser/templates/path_bar.pt
